Fixed `-l` flag for finding live links.
Takes list of URLs and asychrnously calls HEAD request on each
link in the list and test status code for a 200 response. If the
response is not 200 or takes longer than 8 seconds, the link is
declared dead. Also switched from urllib.request to requests for
not only simplicity but thread-safety also.