Fixed -e flag
To use dash -e flag, pass url name like www.url.com
and it will try to establish an https connection than http if
the secure conneciton fails. If both fail then an error message is
printed. If the connection is successful than we search for the
onion domain name and the others that were passed with the flag.