Added Go version of getweblinks getLinks function
Created a Go function that is a replica of getLinks in getweblinks,
except much faster. The go_modules folder is composed of a shared object
file, a header file, the go source code and a python file that is
dynamically linking the shared object.
To it out, just go to the go_modules folder and run `python3
getLinks.py`